+++ sqld/src/replication/replica/mod.rs
@@ -1,5 +1,6 @@
 mod error;
 mod hook;
+mod injector;
 mod meta;
 mod replicator;
 mod snapshot;
